Fixing clang compiling issues

+++ source/Tools/src/Tools.cpp
@@ -114,13 +114,13 @@ auto linearInterpolation(std::pair<int, double> a, std::pair<int, double> b, dou
   std::vector<double> d(n);
   std::vector<double> h(n);
 
-  for (int i = 0; i < n; ++i) {
+  for (size_t i = 0; i < n; ++i) {
     h[i] = x[i + 1] - x[i];
   }
 
   std::vector<double> alpha;
   alpha.push_back(0);
-  for (int i = 1; i < n; ++i) {
+  for (size_t i = 1; i < n; ++i) {
     alpha.push_back(3 * (a[i + 1] - a[i]) / h[i] - 3 * (a[i] - a[i - 1]) / h[i - 1]);
   }
   std::vector<double> c(n + 1);
@@ -131,7 +131,7 @@ auto linearInterpolation(std::pair<int, double> a, std::pair<int, double> b, dou
   mu[0] = 0;
   z[0] = 0;
 
-  for (int i = 1; i < n; ++i) {
+  for (size_t i = 1; i < n; ++i) {
     l[i] = 2 * (x[i + 1] - x[i - 1]) - h[i - 1] * mu[i - 1];
     mu[i] = h[i] / l[i];
     z[i] = (alpha[i] - h[i - 1] * z[i - 1]) / l[i];
